# docs/api.swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
...
    "title": "Hydra OAuth2 \u0026 OpenID Connect Server",
    "contact": {
      "name": "ORY",
      "url": "https://www.ory.am",
      "email": "hi@ory.am"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://github.com/ory/hydra/blob/master/LICENSE"
    },
    "version": "Latest"
  },
  "basePath": "/",
  "paths": {
...
    "/policies": {
      "get": {
        "description": "The subject making the request needs to be assigned to a policy containing:\n\n```\n{\n\"resources\": [\"rn:hydra:policies\"],\n\"actions\": [\"list\"],\n\"effect\": \"allow\"\n}\n```",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "policy"
        ],
        "summary": "List Access Control Policies",
        "operationId": "listPolicies",
        "security": [
          {
            "oauth2": [
              "hydra.policies"
            ]
          }
        ],
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "description": "The offset from where to start looking.",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "description": "The maximum amount of policies returned.",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/policyList"
          },
          "401": {
            "$ref": "#/responses/genericError"
          },
          "403": {
            "$ref": "#/responses/genericError"
          },
          "500": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "description": "The subject making the request needs to be assigned to a policy containing:\n\n```\n{\n\"resources\": [\"rn:hydra:policies\"],\n\"actions\": [\"create\"],\n\"effect\": \"allow\"\n}\n```",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "policy"
        ],
        "summary": "Create an Access Control Policy",
        "operationId": "createPolicy",
        "security": [
          {
            "oauth2": [
              "hydra.policies"
            ]
          }
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/policy"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "policy",
            "schema": {
              "$ref": "#/definitions/policy"
            }
          },
          "401": {
            "$ref": "#/responses/genericError"
          },
          "403": {
            "$ref": "#/responses/genericError"
          },
          "500": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/policies/{id}": {
      "get": {
        "description": "The subject making the request needs to be assigned to a policy containing:\n\n```\n{\n\"resources\": [\"rn:hydra:policies:\u003cid\u003e\"],\n\"actions\": [\"get\"],\n\"effect\": \"allow\"\n}\n```",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "policy"
        ],
        "summary": "Get an Access Control Policy",
        "operationId": "getPolicy",
        "security": [
          {
            "oauth2": [
              "hydra.policies"
            ]
          }
        ],
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The id of the policy.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "policy",
            "schema": {
              "$ref": "#/definitions/policy"
            }
          },
          "401": {
            "$ref": "#/responses/genericError"
          },
          "403": {
            "$ref": "#/responses/genericError"
          },
          "500": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "description": "The subject making the request needs to be assigned to a policy containing:\n\n```\n{\n\"resources\": [\"rn:hydra:policies\"],\n\"actions\": [\"update\"],\n\"effect\": \"allow\"\n}\n```",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "policy"
        ],
        "summary": "Update an Access Control Polic",
        "operationId": "updatePolicy",
        "security": [
          {
            "oauth2": [
              "hydra.policies"
            ]
          }
        ],
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The id of the policy.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/policy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "policy",
            "schema": {
              "$ref": "#/definitions/policy"
            }
          },
          "401": {
            "$ref": "#/responses/genericError"
          },
          "403": {
            "$ref": "#/responses/genericError"
          },
          "500": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "description": "The subject making the request needs to be assigned to a policy containing:\n\n```\n{\n\"resources\": [\"rn:hydra:policies:\u003cid\u003e\"],\n\"actions\": [\"delete\"],\n\"effect\": \"allow\"\n}\n```",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "policy"
        ],
        "summary": "Delete an Access Control Policy",
        "operationId": "deletePolicy",
        "security": [
          {
            "oauth2": [
              "hydra.policies"
            ]
          }
        ],
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The id of the policy.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/emptyResponse"
          },
          "401": {
            "$ref": "#/responses/genericError"
          },
          "403": {
            "$ref": "#/responses/genericError"
          },
          "500": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "policy": {
      "type": "object",
      "properties": {
        "actions": {
          "description": "Actions impacted by the policy.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Actions"
        },
        "conditions": {
          "description": "Conditions under which the policy is active.",
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "properties": {
              "options": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                },
                "x-go-name": "Options"
              },
              "type": {
                "type": "string",
                "x-go-name": "Type"
              }
            }
          },
          "x-go-name": "Conditions"
        },
        "description": {
          "description": "Description of the policy.",
          "type": "string",
          "x-go-name": "Description"
        },
        "effect": {
          "description": "Effect of the policy",
          "type": "string",
          "x-go-name": "Effect"
        },
        "id": {
          "description": "ID of the policy.",
          "type": "string",
          "x-go-name": "ID"
        },
        "resources": {
          "description": "Resources impacted by the policy.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Resources"
        },
        "subjects": {
          "description": "Subjects impacted by the policy.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Subjects"
        }
      },
      "x-go-name": "swaggerPolicy",
      "x-go-package": "github.com/ory/hydra/policy"
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "An empty response"
    },
...
    "genericError": {
      "description": "The standard error format",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "x-go-name": "Details"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "reason": {
            "type": "string",
            "x-go-name": "Reason"
          },
          "request": {
            "type": "string",
            "x-go-name": "Request"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "policyList": {
      "description": "A policy",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/policy"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    },
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://your-hydra-instance.com/oauth2/auth",
      "tokenUrl": "https://your-hydra-instance.com/oauth2/token",
      "scopes": {
        "hydra.clients": "A scope required to manage OAuth 2.0 Clients",
        "hydra.consent": "A scope required to fetch and modify consent requests",
        "hydra.groups": "A scope required to manage warden groups",
        "hydra.health": "A scope required to get health information",
        "hydra.keys.create": "A scope required to create JSON Web Keys",
        "hydra.keys.delete": "A scope required to delete JSON Web Keys",
        "hydra.keys.get": "A scope required to fetch JSON Web Keys",
        "hydra.keys.update": "A scope required to get JSON Web Keys",
        "hydra.policies": "A scope required to manage access control policies",
        "hydra.warden": "A scope required to make access control inquiries",
        "offline": "A scope required when requesting refresh tokens",
        "openid": "Request an OpenID Connect ID Token"
      }
    }
  },
...
}